// src/features/wallet/models/WalletGlobalState.ts
export enum WalletInitStateType {
  IDLE = 'IDLE',
  INIT_REQUESTED = 'INIT_REQUESTED',
  NOT_SUPPORTED = 'NOT_SUPPORTED',
  INIT_FAILED = 'INIT_FAILED',
  INITIALIZED = 'INITIALIZED',
}

export enum WalletAccountStateType {
  IDLE = 'IDLE',
  ACCOUNT_REQUESTED = 'ACCOUNT_REQUESTED',
  ACCOUNT_DETECTION_FAILED = 'ACCOUNT_DETECTION_FAILED',
  LOCKED = 'LOCKED',
  UNLOCK_REQUESTED = 'UNLOCK_REQUESTED',
  UNLOCK_REJECTED = 'UNLOCK_REJECTED',
  UNLOCK_FAILED = 'UNLOCK_FAILED',
  ACCOUNT_LOADED = 'ACCOUNT_LOADED',
}

export enum WalletNetworkStateType {
  IDLE = 'IDLE',
  NETWORK_REQUESTED = 'NETWORK_REQUESTED',
  NETWORK_DETECTION_FAILED = 'NETWORK_DETECTION_FAILED',
  WRONG_NETWORK = 'WRONG_NETWORK',
  NETWORK_SWITCH_REQUESTED = 'NETWORK_SWITCH_REQUESTED',
  NETWORK_SWITCH_REJECTED = 'NETWORK_SWITCH_REJECTED',
  NETWORK_SWITCH_FAILED = 'NETWORK_SWITCH_FAILED',
  NETWORK_LOADED = 'NETWORK_LOADED',
}

export enum WalletSignStateType {
  IDLE = 'IDLE',
  NOT_SIGNED = 'NOT_SIGNED',
  SIGN_REQUESTED = 'SIGN_REQUESTED',
  SIGN_REJECTED = 'SIGN_REJECTED',
  SIGN_TIMED_OUT = 'SIGN_TIMED_OUT',
  SIGN_FAILED = 'SIGN_FAILED',
  SIGNED = 'SIGNED',
}

export enum WalletStateType {
  NOT_INITIALIZED = 'NOT_INITIALIZED',
  CHECKING_WALLET = 'CHECKING_WALLET',
  CHECKING_ACCOUNT = 'CHECKING_ACCOUNT',
  CHECKING_NETWORK = 'CHECKING_NETWORK',
  CHECKING_SIGN = 'CHECKING_SIGN',
  AUTHENTICATED = 'AUTHENTICATED',
}

export type WalletGlobalStateType = {
  state: WalletStateType;
  initState: WalletInitStateType;
  accountState: WalletAccountStateType;
  networkState: WalletNetworkStateType;
  signState: WalletSignStateType;
};